move types in single file
