+++ packages/marketing-cloud-logo/src/types.ts
