--- packages/marketing-cloud-logo/src/types/index.ts
@@ -1,7 +1,0 @@
-export type {
-  FillColor,
-  FillStyle,
-  HexColorCode,
-  LogoType,
-  Size
-} from './commonProps';
